test(hwpx): 헤더 매퍼 테스트를 매핑·진단으로 나눈다
한 클래스가 SwiftLint type_body_length error 임계(400줄)에 닿아 진단·강등
계열 테스트를 더 넣을 수 없었다. 매핑 결과 검증은 그대로 두고 진단으로
강등되는지 보는 6건을 HwpxHeaderDiagnosticsTests로 옮기고, 두 파일이 함께
쓰는 headerXML·mapHeader를 HwpxHeaderFixture로 승격한다.

테스트 내용은 옮기기만 했고 단언은 그대로다.
